修改了商品删除时未对不存在的商品判定的问题

+++ backend/src/main/java/com/example/tomatomall/service/serviceImpl/ProductSreviceImpl.java
@@ -144,6 +144,10 @@ public class ProductSreviceImpl implements ProductService {
 
     @Override
     public String delete(Integer id) {
+        Optional<Product> thisProduct=productRepository.findById(id);
+        if(!thisProduct.isPresent()){
+            return "商品不存在";
+        }
         productRepository.deleteById(id);
 
         List<Specification> specifications=specificationRepository.findByProductId(id);
